src/ghostbusters_api/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# GCP Services
google-cloud-firestore==2.11.1
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.16.4
functions-framework==3.4.0

# HTTP client
requests==2.31.0

# Basic dependencies
typing-extensions==4.8.0 